[PORT FROM R4.1] Disable video effects feature for yukkabeach and lexington
BZ: 66003

Issue:
To enable the video feature: video effects feature, it needs more
video buffers in the camera HAL.
The video effects is not one POR for yukkabeach and lexington.
If we can disable this feature, we can reduce the video buffers number
in the camera HAL.
disable the feature in the video side.

Solution:
set the INPORT_ACTUAL_BUFFER_COUNT to 2 if we don't need the video effects feature.

Depends on:
None.

Change-Id: I799c915047cb4f942244b6ce810c85f777f945d7
Orig-Change-Id: Ib3ca3e3bd259a5e0fa4b1e3e15257a8ccba6bdba
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/78643
Reviewed-by: Chang, Ying <ying.chang@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/87002
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>